+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- Canonicalise `anyOf` special cases when all subschemas have only the `type` keyword
+
 #### 0.18.0 - 2020-09-10
 - Performance improvements from careful caching (#62)
 - Use a validator that corresponds to the input schema draft version (#66)